Servicio y controladores de corral

+++ BackendSpringboot/src/main/java/RIwi/BackendSpringboot/model/repositories/AnimalRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface AnimalRepository extends JpaRepository<AnimalEntity,Integer> {
-    List<AnimalEntity> findAllById(Iterable<Integer> ids);
+public interface AnimalRepository extends JpaRepository<AnimalEntity,Long> {
+    List<AnimalEntity> findAllById(Iterable<Long> ids);
     List<AnimalEntity> findByCorralId(Long corralId);
     List<AnimalEntity> findByDangerous(boolean dangerous);
 }
